+++ lib/src/background.dart
@@ -1,3 +1,5 @@
+library background;
+
 import 'dart:math';
 import 'dart:html';