+++ lib/src/movements.dart
@@ -1,3 +1,4 @@
+library movements;
 import 'dart:math';
 
 /// The Movements class provides pairs of values for setting movements.