+++ lib/src/renderer.dart
@@ -1,3 +1,5 @@
+library renderer;
+
 import 'dart:html';
 import 'background.dart';
 import 'spritegroup.dart';